requirements.txt
pandas
numpy
scikit-learn
notebook
pyYAML
matplotlib
imblearn
django
-e .